Don't run CTR-DRBG in Kyber tests.
We didn't need to do this, but we ran CTR-DRBG to generate values that
are also in the test vectors and checked that they matched. We will very
likely switch to using a derivation function in CTR-DRBG in the future,
due to NIST's changing requirements, and this will stop working then.

Change-Id: Id7fa678ad602d6a7387158ad84c971aa2121c73f
Reviewed-on: https://boringssl-review.googlesource.com/c/boringssl/+/79767
Commit-Queue: Adam Langley <agl@google.com>
Reviewed-by: David Benjamin <davidben@google.com>