Supress the run_tests target when we're a subdirectory
gRPC imports us as a subdirectory and CMake has no namespacing to speak
of for that pattern.
https://gitlab.kitware.com/cmake/cmake/-/issues/16414
https://gitlab.kitware.com/cmake/cmake/-/issues/22687

The run_tests target is not great in that model, both because it's
unnamespaced, and because it reportedly conflicts with CTest's RUN_TESTS
target, when generating MSVC projects because it's case-insensitive.

Probably what we actually want is to register our tests with CTest and
somehow convince CTest to behave reasonably, but for now let's just
suppress it.

Also namespace the all_tests helper target to avoid that leaking
unhelpfully. Though really it's meant to be private helper target and
not exported anywhere.

We probably should similarly gate all our warning flags and other
dev-specific defaults on PROJECT_IS_TOP_LEVEL, but I'll leave that for
later.

Bug: 428112180
Change-Id: Ib7c09c3aa083e373c97a9a5a903459b1eb7b56f4
Reviewed-on: https://boringssl-review.googlesource.com/c/boringssl/+/80188
Auto-Submit: David Benjamin <davidben@google.com>
Reviewed-by: Adam Langley <agl@google.com>
Commit-Queue: Adam Langley <agl@google.com>