Parse X509_ALGOR without depending on the templates
We do still need to give X509_ALGOR an ASN1_ITEM, both because the
containing types have not been migrated, and because there are external
callers that use <openssl/asn1t.h> and embed X509_ALGOR.
https://crbug.com/42290417#comment2 has a list of those.

This is a bit more boilerplate than I'd like. Some comes from just
needing to define a bunch of functions. Some comes from not being able
to use C++ ctors and dtors, which is a bit hard to avoid because
X509_ALGOR is actually a public struct. Some is because defining an
ASN1_ITEM involves a bit of code. I've hidden that last one behind a
macro.

(In the long run, I'd like to make the ASN1_ITEM callbacks
CBS/CBB-based, but that's a bit tied up with rewriting tasn_enc.cc and
tasn_dec.cc, which is in turn tied up with the X509_NAME parse callback,
which depends on ASN1_item_ex_d2i and ASN1_item_ex_i2d.)

Bug: 42290417
Change-Id: Ib0bfc9c942aa6bf784cf15af6b3747c9fc9a88f5
Reviewed-on: https://boringssl-review.googlesource.com/c/boringssl/+/81777
Commit-Queue: David Benjamin <davidben@google.com>
Reviewed-by: Adam Langley <agl@google.com>