Test verifying signatures over "unusual" TBSCertificates
Test that, signatures over unusual TBSCertificates are verified
correctly. This tests that encoding is correctly round-tripped through
the parser to the verifier.

In principle, this should never happen because a DER parser will only
accept the canonical encoding of an object. However, it is possible for
encoding to not round-trip if we accept any BER inputs, or our in-memory
representation does not capture the full range of abstract
TBSCertificate values.

X509 objects cache the encoded TBSCertificate, so all encoding
variations should be captured. This test tries to exercise the cache's
effects on signature verification. In reality, the cache is barely
load-bearing. We now reject most non-DER inputs, and X509_NAME also
saves its encoding. Still, the test ensures this remains the case.

Amusingly, you're actually formally *not* supposed to do this. An X.509
signature is defined to use the DER encoding of the TBSCertificate. An
ASN.1 type is independent of the particular encoding used. If you parsed
a BER (or XER!) X.509 Certificate, you're meant to canonicalize the
encoding back to DER to verify it. As far as I know, no one does it that
way.

Change-Id: I67c4932ec87d1f6fedfc8c711cf0f7837759947e
Reviewed-on: https://boringssl-review.googlesource.com/c/boringssl/+/81768
Commit-Queue: David Benjamin <davidben@google.com>
Reviewed-by: Adam Langley <agl@google.com>
Auto-Submit: David Benjamin <davidben@google.com>