Test the implicit cleanup feature of X509_sign_ctx, etc.
The *_sign_ctx APIs implicitly call EVP_MD_CTX_cleanup on return, on
both success and failure. Some callers rely on this to avoid a memory
leak. An earlier iteration of this patch set accidentally broke this.
Add a test so we catch this.

Change-Id: I241b9161606f91adfbc37064304b815145703659
Reviewed-on: https://boringssl-review.googlesource.com/c/boringssl/+/81807
Commit-Queue: David Benjamin <davidben@google.com>
Reviewed-by: Adam Langley <agl@google.com>