Always populate supported_group_list
SSL_CONFIG and SSL_CTX have a supported_group_list field which stores
the ordered list of group IDs of groups supported for key exchange.
These can be set by the caller via several setters (that either take
group IDs, NIDs, or a string list), but there is a default list that is
used if the caller hasn't configured the supported groups.

Currently the default is evaluated by a getter, tls1_get_grouplist(),
but it is more convenient if the supported_group_list field can be
treated as a source of truth, i.e. if it contains the applicable groups
at all times.

This CL eliminates the getter, and instead populates
supported_group_list with the default groups in SSL_CTX_new.

The behavior of the various setters is preserved. The group ID and NID
based setters now special-case an empty input to restore the default
list. The string list based setter never accepted an empty string, so
no changes are made to that. The public API documentation now mentions
these behaviors explicitly.

Update-Note: There should not be any behavior change. The documentation
is updated to clarify what was existing behavior and is still the
behavior of providing an empty list when setting supported groups.

Change-Id: Iced20f80deb79fbd4b8bd560a00bf395e9b0e86b
Reviewed-on: https://boringssl-review.googlesource.com/c/boringssl/+/81529
Reviewed-by: David Benjamin <davidben@google.com>
Commit-Queue: Lily Chen <chlily@google.com>