Allow SSL_HANDSHAKE::key_shares to vary in size
This refactors hs->key_shares to a vector rather than a fixed array of
size 2. There is at most a key share for every implemented NamedGroup,
so the size is capped at the number of named groups. This is needed for
allowing the caller to explicitly specify client key shares. There's no
behavior change intended in this CL. We still use at most 2 key shares
(by default; that will change in the next CL).

Change-Id: I0eeb218c59ec1f0e8e062aeb5f833102c40d7bf6
Bug: 437414371
Reviewed-on: https://boringssl-review.googlesource.com/c/boringssl/+/81487
Reviewed-by: David Benjamin <davidben@google.com>
Commit-Queue: Lily Chen <chlily@google.com>