Make make_unusual_tbs.go test data idempotent
If the signature is already valid, don't bother regenerating the test
data. It's somewhat irksome that there isn't a crypto.Verifier to mirror
crypto.Signer, but ah well.

I've only added ECDSA support for now, but we can extend it later if we
want to. (In principle we could use this to generate more of our test
certs, not just the "unusual TBS" ones. This is automating the annoying
step of
https://github.com/google/der-ascii/blob/main/samples/certificates.md )

Change-Id: I0c3dac566a9a0f808b6e91de4d521f794d0503a8
Reviewed-on: https://boringssl-review.googlesource.com/c/boringssl/+/82007
Reviewed-by: Lily Chen <chlily@google.com>
Commit-Queue: Lily Chen <chlily@google.com>
Auto-Submit: David Benjamin <davidben@google.com>