Add API for configuring client key shares
This change introduces a new API to allow callers to configure the exact
set of key shares to be sent in a client's key_share extension. If the
supported groups for a connection are modified, any previously-selected
key shares are cleared if they are no longer compatible. Clients are
allowed to configure an empty list of key shares, which results in
always taking a round-trip for HelloRetryRequest.

Bug: 437414371
Change-Id: Ibd2a9b217fa3c746dec194a027d3ec45a81bb578
Reviewed-on: https://boringssl-review.googlesource.com/c/boringssl/+/81567
Commit-Queue: Lily Chen <chlily@google.com>
Reviewed-by: David Benjamin <davidben@google.com>