Upgrade the opportunistic CRYPTO_sysrand calls to normal ones
We do not believe this is needed anymore. See
go/boringssl-and-os-entropy (internal).

After this change, we still use RDRAND for the seed and the OS for the
personalization string. There's no reason for this now. Rather, this is
meant to be a more easily reverted change. I've left some TODOs to clean
this up, once this sticks.

Update-Note: Builds with BORINGSSL_FIPS will no longer attempt to work
around broken OS entropy with RDRAND. The OS is expected to be
functional.

Bug: 446280903
Change-Id: I60bf9c4c0c1268d76b26301b41097179de453120
Reviewed-on: https://boringssl-review.googlesource.com/c/boringssl/+/82108
Reviewed-by: Adam Langley <agl@google.com>
Auto-Submit: David Benjamin <davidben@google.com>
Commit-Queue: Adam Langley <agl@google.com>