Make the ASN1_TYPE-level type take precedence over the ASN1_STRING one
It is not very difficult to get an ASN1_TYPE that is self-inconsistent,
where the ASN1_TYPE-level type says one thing and the ASN1_STRING-level
type says another.

In particular, a not entirely unreasonable caller pattern used in
Android does this when filling in X509_ALGOR parameters. The old
ASN1_TYPE encoder took the ASN1_TYPE type, but the new one takes the
ASN1_STRING type.

Match the old behavior. We probably should also make ASN1_TYPE_set0
internally fix the type of the ASN1_STRING, but I haven't done that
here. I think, either way, we should restore the old behavior of using
the outer type. With the mess of exposed types and ambiguous APIs in
OpenSSL, it's not too hard to manually construct something invalid here.

Bug: 446993031
Change-Id: Idd5c0356769d8ff9a034222a9ffc107ce7c6f02e
Reviewed-on: https://boringssl-review.googlesource.com/c/boringssl/+/82187
Reviewed-by: Lily Chen <chlily@google.com>
Commit-Queue: David Benjamin <davidben@google.com>