Put the legacy cipher constants back in tls1.h and ssl3.h
QUICHE expects TLS1_CK_* to be defined in tls1.h, not ssl.h. For now,
just define them in tls1.h as before. Ideally we'd define them in terms
of the other constants, but we'd then either need to move SSL_CIPHER_*
out of ssl.h or merge tls1.h and ssl3.h into ssl.h. The latter is
probably a better direction, but is a much larger change, so start with
this.

This should unbreak the Chromium roll.

Change-Id: Ic403bef7043028d30aa9916b403ebb3721734f32
Reviewed-on: https://boringssl-review.googlesource.com/c/boringssl/+/82387
Auto-Submit: David Benjamin <davidben@google.com>
Reviewed-by: Lily Chen <chlily@google.com>
Commit-Queue: Lily Chen <chlily@google.com>