Add API for caller to hint server's preferred key shares
This change introduces a new mechanism for clients to provide a "hint"
about which key exchange groups the server is likely to support. This is
meant to be used for draft-ietf-tls-key-share-prediction.

We will try to make a prediction based on the server's hint, and if a
group was predicted, it is sent as the key_share, overriding any
explicit key shares configured by the caller.

Bug: 437414371
Change-Id: Ie98d2a63e5bfb4f10683ca4f1f77e5439b39e256
Reviewed-on: https://boringssl-review.googlesource.com/c/boringssl/+/81588
Reviewed-by: David Benjamin <davidben@google.com>
Commit-Queue: Lily Chen <chlily@google.com>