Also accept incorrectly-encoded X509 v1 version fields again
This reverts another part of
https://boringssl-review.googlesource.com/c/boringssl/+/82087

Sadly this hit another real-world input. We'll need to use our new
parser to wire in a X509_parse_with_options or something of the sort, so
that the impacted code can pass in arguments to reject a handful of
non-certificate variations.

Bug: b:449478579, 42290225
Change-Id: I1e39c8b6ed0ab454bf61c85edffe6631e8c7c3df
Reviewed-on: https://boringssl-review.googlesource.com/c/boringssl/+/82747
Auto-Submit: David Benjamin <davidben@google.com>
Commit-Queue: Adam Langley <agl@google.com>
Reviewed-by: Adam Langley <agl@google.com>